requirements.txt
langchain
sentence-transformers
qdrant-client
python-docx
tabula-py
pytesseract
